Support style invalidation of ':has()' in rightmost compound
This is the initial CL of supporting style invalidation with ':has()'.

To minimize complexity, we will focus on the following limited usage.
- Supports ':has()' in a rightmost, top-level compound
- Supports ':has()' argument starts with child or descendant combinator
- Only allow attribute or elemental selectors in argument

So the following will be ignored for now.
- Ignore ':has()' in a non-rightmost compound
- Ignore ':has()' in a logical combinations
- Ignore ':has()' argument starts with '~' or '+'
- Ignore all pseudos in ':has()' argument

To support this case, for a change on an element, we need to invalidate
ancestors.

Invalidate every ancestors for every mutations on every element is very
bad for performance.

So we need a way to skip these.
- irrelevant element
- irrelevant ancestor
- irrelevant mutation

To skip the 'irrelevant element' and 'irrelevant ancestor', this CL
introduces two flags in the ComputedStyle.
- AffectingHas   : Indicate that a mutation on the element can affect
                   ':has()' state.
- AffectedByHas  : Indicate that the element can be affected by ':has()'
                   state change

During style recalculation, 'AffectedByHas' flag will be set when
SelectorChecker tries to match any ':has()' selector on an element.
For every descendants of the element, the 'AffectedByHas' flag will
be changed to 'AffectingHas' flag and stored.

<style> .a:has(.b) { color: red }</style>
<div>
  <div class="a">    <!-- AffectedByHas  -->
    <div>            <!-- AffectingHas  -->
      <div>          <!-- AffectingHas  -->
      </div>
    </div>
  </div>
  <div>
    <div>
    </div>
  </div>
</div>

Only for a mutation on the element with 'AffectingHas', style engine
will walk up ancestors. And it invalidate style only when an ancestor
has 'AffectedByHas'

Still we have invalidation for irrelevant mutation and invalidation on
irrelevant ancestors. We will handle those by separate CLs.

Bug: 669058
Change-Id: Ia7e607e01f9a070c0d69d89679b99efa0a104797