Don't skip body background when not propagated
Body background propagation to viewport is stopped when html root or
body is contained, which means we should paint the background for the
body element itself when that happens.

The wpt test did test we didn't propagate the background to the
viewport, but not that we painted it for body. Modify the code and the
test to make sure the body background is painted.

Bug: 1266731
Change-Id: I405ac1dde17fc059f5145ba295f0d3e8e1ca1ba3